aim9: Use exp_continue instead of while { expect ... } pattern
expect complains with 'expect' executed in while loop. As it looks,
the while loop is used to wait for all tests to terminate. aim9 will
then output "disk_rr". Unfornatly, every tests also sends EOF so we
can't wait for EOF, since we would wait only for the first to
terminate. Instead use exp_continue for all lines which don't have
"disk_rr" in it.

Signed-off-by: Daniel Wagner <dwagner@suse.de>